renaming the doc-to-pdf script
